refactor: 총 자산 금액 totalAmount을 Long->BigDecimal로 변환

+++ backend/src/main/java/com/peekport/dto/GoalAccountResponse.java
@@ -13,7 +13,7 @@ import java.math.BigDecimal;
 public class GoalAccountResponse {
     private Long id;
     private String name;
-    private Long totalAmount;
+    private BigDecimal totalAmount;
     private Long targetAmount;
     private BigDecimal cash;
     private PortfolioType portfolioType;
